Fixed a number of random FindBugs issues.
git-svn-id: https://svn.apache.org/repos/asf/incubator/jspwiki/trunk@666848 13f79535-47bb-0310-9956-ffa450edef68

+++ src/com/ecyrd/jspwiki/auth/login/CookieAuthenticationLoginModule.java
@@ -329,7 +329,7 @@ public class CookieAuthenticationLoginModule extends AbstractLoginModule
 
         File[] files = cookieDir.listFiles();
 
-        long obsoleteDateLimit = System.currentTimeMillis() - (days+1) * 24 * 60 * 60 * 1000L;
+        long obsoleteDateLimit = System.currentTimeMillis() - ((long)days+1) * 24 * 60 * 60 * 1000L;
 
         int  deleteCount = 0;
